Sales Report
A Power BI dashboard for an e-commerce business. I added an average line that shows how a certain day's results compare with a normal day and how to use Power BI to perform a deep dive into the underlying data and tell us what is causing an increase or decrease.